Cambiar las propiedades CurrentPage y TotalPages a long por coherencia

+++ domain/Core/Page.cs
@@ -4,8 +4,8 @@
     {
         public List<T> Items { get; set; } = new List<T>();
         public required long TotalCount { get; set; }
-        public required int CurrentPage { get; set; }
+        public required long CurrentPage { get; set; }
         public required int PageSize { get; set; }
-        public int TotalPages => (int)Math.Ceiling((double)TotalCount / PageSize);
+        public long TotalPages => (long)Math.Ceiling((double)TotalCount / PageSize);
     }
 }
